refactor: 이미지 변경 (#43)
- 마이 페이지에서 동일한 이미지를 사용하기 위해 이미지 변경
